+++ QuickStackStore/Package/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## 1.3.5
+- Fixed area quick stacking potentially causing issues when affecting a container at the edge of the loading range due to a very high range setting
+- Added Swedish translation
+- Updated Polish translation
 ## 1.3.4
 - Added compatibility for AzuExtendedPlayerInventory
 - All trashing features are now disabled when using RecyclePlus